requirements.txt
flask
python-dotenv
groq
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
